Made changes requested in pull request #48 - issue #28
