--- requirements.txt
@@ -1,1 +1,0 @@
-mezzanine >= 4.2.0
